requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
stripe==7.4.0
anthropic==0.18.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.23.1
alembic==1.13.0
python-multipart==0.0.6
aiofiles==23.2.1
pypdf==4.1.0
python-docx==1.1.0
email-validator==2.1.0
spacy>=3.7.0
sentence-transformers==2.3.1
scikit-learn==1.4.0
faiss-cpu==1.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
tenacity==8.2.3
pandas==2.1.4
openpyxl==3.1.2
